fixed data types warnings

+++ writeInROM.cpp
@@ -12,51 +12,51 @@ void writeInROM(struct MyNavigator* myNavigator)
 	{
 
 		// Широта :
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lat.decodedPos.value);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lat.decodedPos.value);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lat.filteredPos.allFilteredValues[0]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lat.filteredPos.allFilteredValues[0]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lat.filteredPos.allFilteredValues[1]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lat.filteredPos.allFilteredValues[1]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lat.filteredPos.allFilteredValues[2]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lat.filteredPos.allFilteredValues[2]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lat.filteredPos.allFilteredValues[3]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lat.filteredPos.allFilteredValues[3]);
 		myNavigator->writeInFlash.adress += 8;
 
 		// Долгота :
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lon.decodedPos.value);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lon.decodedPos.value);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lon.filteredPos.allFilteredValues[0]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lon.filteredPos.allFilteredValues[0]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lon.filteredPos.allFilteredValues[1]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lon.filteredPos.allFilteredValues[1]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lon.filteredPos.allFilteredValues[2]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lon.filteredPos.allFilteredValues[2]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.lon.filteredPos.allFilteredValues[3]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.lon.filteredPos.allFilteredValues[3]);
 		myNavigator->writeInFlash.adress += 8;
 
 		// Высота :
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.alt.decodedPos.value);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.alt.decodedPos.value);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.alt.filteredPos.allFilteredValues[0]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.alt.filteredPos.allFilteredValues[0]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.alt.filteredPos.allFilteredValues[1]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.alt.filteredPos.allFilteredValues[1]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.alt.filteredPos.allFilteredValues[2]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.alt.filteredPos.allFilteredValues[2]);
 		myNavigator->writeInFlash.adress += 8;
 
-		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (unsigned long long int)myNavigator->coordinates.alt.filteredPos.allFilteredValues[3]);
+		(*myNavigator->writeInFlash.flashFunc)(0x00000003U, myNavigator->writeInFlash.adress, (uint64_t)myNavigator->coordinates.alt.filteredPos.allFilteredValues[3]);
 		myNavigator->writeInFlash.adress += 8;
 
 		// Время с прошлого сообщения и время выполнения функции myNavigator :
